Devoir 1

+++ basics/src/main/java/be/ecam/basics/exercises/SensorWindow.java
@@ -3,10 +3,10 @@ package be.ecam.basics.exercises;
 public class SensorWindow {
     public static int average(int[] values) {
         if (values == null || values.length == 0) throw new IllegalArgumentException("values");
-        int sum = 0;
+        long sum = 0; // utiliser long pour éviter l'overflow
         for (int v : values) {
             sum += v;
         }
-        return sum / values.length;
+        return (int)(sum / values.length); // division entière, puis cast en int
     }
 }
